move test_xxx to basic test class
